requirements.txt
fastapi>=0.100.0
httpx[socks]>=0.24.0
openai>=1.0.0
pydantic>=2.0.0
pydantic_settings>=2.0.0
requests>=2.28.0
starlette>=0.27.0
uvicorn[standard]>=0.20.0
google-genai>=0.3.0
jinja2>=3.1.0
python-multipart>=0.0.6
cryptography>=3.4.8
pymysql>=1.0.0
sqlalchemy>=2.0.0
aiomysql>=0.1.1
aiosqlite>=0.19.0
databases[sqlite]>=0.7.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
packaging>=21.0
